mk1/requirements.txt
gradio>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
ultralytics>=8.0.0
transformers>=4.30.0
pyyaml>=6.0.0
deep-sort-realtime>=1.3.0
scikit-image>=0.20.0
tqdm>=4.65.0 